==> 01-scraping_data.ipynb <==
## 01 - Scraping Data
This notebook provides the code to scrape the html `https://www.congress.gov/search` website for all nominations for the senate judiciary committee from the 107th to the 117th congress.

###Import scraping libraries
from bs4 import BeautifulSoup
import pandas as pd
import re
import requests
from playwright.async_api import async_playwright

list = [1,2,3,4,5,6,7,8,9,10,11,12]

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless = False)
page = await browser.new_page()

for i in list:

    await page.goto(f"https://www.congress.gov/search?q=%7B%22source%22%3A%22nominations%22%2C%22senate-committee%22%3A%22Judiciary%22%2C%22congress%22%3A%5B%22107%22%2C%22108%22%2C%22109%22%2C%22110%22%2C%22111%22%2C%22112%22%2C%22113%22%2C%22114%22%2C%22115%22%2C%22116%22%2C%22117%22%5D%7D&pageSize=250&page={i}")

    await page.content()
    doc = BeautifulSoup(await page.content(), 'html.parser')

    with open (f'html_scrapes/output{i}.html', 'w') as file:
        file.write(str(doc))
