==> 02-scapring_html.ipynb <==
## 02 - Scraping the HTML
This notebook scrapes and clean up the html from notebook 1: `01 - Scraping Data`. 

import re
import requests
from bs4 import BeautifulSoup

list = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]

nominations = []

# Loop through each html-file and extract the nominations.
for i in list:
    with open(f'html_scrapes/output{i}.html', 'r') as f:

        contents = f.read()

        doc = BeautifulSoup(contents, 'lxml')

    #Grab the HTML that contains the nominations. 

    results = doc.select('#main')

    #And then I create a variable (a list) which contains each nomination.

    elements = results[0].select('.expanded')

    for element in elements:
        nomination = {}
        nomination['name'] = element.strong.text
        # And remove "— Department of Justice" or " — The Judiciary" from the names
        nomination['name']= nomination['name'].replace(' — Department of Justice', '')
        nomination['name']= nomination['name'].replace(' — The Judiciary', '')
    
        # And then I grab all the nomination numbers. 
        nomination['nomination_no'] = element.a.next
    
        nomination['congress_no'] = element.strong.previous.strip()
        # And remove the extra – and (year-year)
        nomination['congress_no']= nomination['congress_no'].replace('—', '').strip()
    
    
        # Here I grab the information on the circuit for the nominations for Circuit Court judges. All the rest of the nominations are passed. 
        if "Circuit" in element.find(class_="result-item").text.strip():
            nomination['post'] = element.find(class_="result-item").text.strip()
        else:
            pass
    
        # Here I grab the information on the votes and the day of the confirmation but only for the nominees that have been confirmed by Senate.
        if "Confirmed" in element.find_all(class_="result-item")[-1].text.strip():
            nomination['vote'] = element.find_all(class_="result-item")[-1].text.strip()
        else:
            pass
    
        #print(nomination)
        nominations.append(nomination)

import pandas as pd
df = pd.DataFrame(nominations)

Drop all the rows that where post or vote is NaN. Those are the entries where the nomination isn't for a circuit judge or where the nomination has not been confirmed by Congress. 

df1 = df.dropna(subset=['post','vote']) 

And then I save the data to a csv-file.

df1.to_csv("data/nominations.csv", index=False)
