==> 03-cleaning_data.ipynb <==
## 03 - Cleaning the data

This notebook continues the data cleaning from notebook `02-scapring_html`. It uses regular expressions to parce through the html bits and extract the wanted information.

import pandas as pd
import re

df = pd.read_csv("data/nominations.csv")

# Strip the concress_no column of any text and years. 
df['congress_no'] =  df['congress_no'].apply(lambda x: re.sub(r'th Congress .\d{4}-\d{4}.','', str(x)))

df['circuit'] = df['post'].str.extract('(the \w.+ Circuit)', expand=False).str.strip()

# And strip the column of any additional text so only the circuit names remain.
df['circuit'] = df['circuit'].str.replace('the District of Columbia, to be United States Circuit Judge for ','')
df['circuit'] = df['circuit'].str.replace('the ','')
df['circuit'] = df['circuit'].str.replace(' Circuit','')

df['confirmation_date'] = df['vote'].str.extract('(\d\d\W\d\d\W\d{4})', expand=False).str.strip()
df.head(5)

# I make a list out of the column vote so I can use regex on the information in the column. 
list_vote = df.vote.tolist()

for list in list_vote:
    try:
        sub_regex = r"Latest Action: \d\d\W\d\d\W\d{4}.*(\d\d+) - (\d*).*: (\d+)[.].*"
        df.vote.str.split(sub_regex,expand=True)
    except:
        pass

votes = df.vote.str.split(sub_regex,expand=True)

df["yea_votes"]= votes[1]
df["nay_votes"]= votes[2]
df["record_vote_number"]= votes[3]

I realized that the regex, in the rare cases where a judges was confirmed with 100 yea votes, has cut the number to two digits. Thus, I have to replace the 00 with 100 in the yea column. No judge will ever be confirmed with 00 votes, so it won't mess up the rest of the data. 

df.loc[ df['yea_votes'] == "00", 'yea_votes'] = "100"

data = df.drop(['post', 'vote'], axis=1)

data.to_csv("data/votes.csv", index=False)
