==> 05-merging_data.ipynb <==
## 05 - Merging Data

This notebook merges the two scrapes from `03 - Cleaning Data` and `04 - Party Votes` into one dataframe.

import pandas as pd

df = pd.read_csv("data/voteinfo.csv", na_values=['NaN'])
df1 = pd.read_csv("data/party_votes.csv", na_values=['NaN'])

df['vote_id'] = df['vote_id'].str.replace(r'<NA>', 'nan', regex=True)

df1['vote_id1'] = df1['vote_id']
df1 = df1.drop(['congress_no', 'session','vote_number', 'vote_id'], axis=1)

df2 = pd.concat([df, df1], axis=1)

congress_votes = []

for index, row in df2.iterrows():
    senators = {}
    senators["democrats"] = row.D_yea + row.D_nay + row.D_no_vote
    senators["republicans"] = row.R_yea + row.R_nay + row.R_no_vote
    senators["independents"] = row.I_yea + row.I_nay + row.I_no_vote    
    congress_votes.append(senators)
congress_votes

df3 = pd.DataFrame(congress_votes)

data = pd.concat([df2, df3], axis=1)
cols = list(data.columns.values)

data = data[['name', 
'circuit', 
'confirmation_date', 
'nomination_no',
'congress_no',
'session', 
'democrats',
'republicans', 
'independents', 
'record_vote_number', 
'yea_votes',
'nay_votes', 
'vote_id',
'D_yea',
'D_nay',
'D_no_vote',
'R_yea',
'R_nay',
'R_no_vote',
'I_yea',
'I_nay',
'I_no_vote',
'votes']]

data.head(5)

data.to_csv("data/data.csv", index=False)
